Added getLocation feature to app.
