+++ app/src/main/java/com/example/beetrackerrough/BeeSightingUserInput.kt
@@ -1,42 +1,33 @@
 package com.example.beetrackerrough
 
-import android.annotation.SuppressLint
-import android.content.pm.PackageManager
 import android.os.Bundle
-import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.EditText
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.content.ContextCompat
+import android.widget.Toast
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.google.android.gms.location.FusedLocationProviderClient
-import com.google.android.gms.location.LocationServices
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.OnMapReadyCallback
-import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.Marker
-import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.firestore.GeoPoint
+import java.util.Date
+import java.util.Timer
 import java.util.UUID
+import kotlin.concurrent.schedule
 
 
 private const val TAG = "BEE_SIGHTING_USER_INPUT_FRAGMENT"
-private const val LATITUDE = "latitude"
-private const val LONGITUDE = "longitude"
+private const val LAT = "latitude"
+private const val LONG = "longitude"
+
 /**
  * A simple [Fragment] subclass.
  * Use the [BeeSightingUserInput.newInstance] factory method to
  * create an instance of this fragment.
  */
 class BeeSightingUserInput : Fragment() {
-    private var latitude: Double? = null
-    private var longitude: Double? = null
+    private var userLocation: GeoPoint? = null
+
 
     private var beeUserSightingNumInput: Int = 0
 
@@ -48,130 +39,19 @@ class BeeSightingUserInput : Fragment() {
     private lateinit var getLocationButton: Button
     private lateinit var submitButton: Button
 
-    private var locationPermissionGranted = false
-
-    private var moveMapToUsersLocation = false
-
-    private var fusedLocationProvider: FusedLocationProviderClient? = null
-
-    private var map: GoogleMap? = null
+    private val beeViewModel: BeeViewModel by lazy {
+        ViewModelProvider(requireActivity()).get(BeeViewModel::class.java)
+    }
 
     //private val beeMarkers = mutableListOf<Marker>()
 
     //private var beeList: listOf<Bee>()
 
-
-    private val beeViewModel: BeeViewModel by lazy {
-        ViewModelProvider(requireActivity()).get(BeeViewModel::class.java)
-    }
-
-    private val mapReadyCallback = OnMapReadyCallback { googleMap ->
-        Log.d(TAG, "Google map ready")
-        map = googleMap
-
-        updateMap()
-    }
-
-
-    @SuppressLint("MissingPermission")
-    private fun moveMapToUserLocation() {
-        if (map == null) {
-            return
-        }
-
-        if (locationPermissionGranted) {
-            map?.isMyLocationEnabled = true
-            map?.uiSettings?.isMyLocationButtonEnabled = true
-            map?.uiSettings?.isZoomControlsEnabled = true
-
-            fusedLocationProvider?.lastLocation?.addOnCompleteListener { getLocationTask ->
-                val location = getLocationTask.result
-                if (location != null) {
-                    Log.d(TAG, "User's location $location")
-                    val center = LatLng(location.latitude, location.longitude)
-                    val zoomLevel = 8f
-                    map?.moveCamera(CameraUpdateFactory.newLatLngZoom(center, zoomLevel))
-                    moveMapToUsersLocation = true
-                } else {
-                    showSnackbar(getString(R.string.no_location))
-                }
-            }
-        }
+    private fun toGeoPoint(userLocationLat: Double, userlocationLong: Double) {
 
     }
 
-    private fun updateMap() {
-        if (locationPermissionGranted) {
-            if (!moveMapToUsersLocation) {
-                moveMapToUserLocation()
-            }
-        }
-    }
 
-    private fun requestLocationPermission() {
-        //if (isAdded()) {
-            if (ContextCompat.checkSelfPermission(
-                    requireActivity(),
-                    android.Manifest.permission.ACCESS_FINE_LOCATION
-                ) == PackageManager.PERMISSION_GRANTED
-            ) {
-                locationPermissionGranted = true
-                Log.d(TAG, "permission already granted")
-                updateMap()
-                fusedLocationProvider =
-                    LocationServices.getFusedLocationProviderClient(requireActivity())
-            } else {
-                val requestLocationPermissionLauncher =
-                    registerForActivityResult(ActivityResultContracts.RequestPermission()) { granted ->
-                        if (granted) {
-                            Log.d(TAG, "User granted permission")
-                            locationPermissionGranted = true
-                            fusedLocationProvider =
-                                LocationServices.getFusedLocationProviderClient(requireActivity())
-                        } else {
-                            Log.d(TAG, "User did not grant permission")
-                            locationPermissionGranted = false
-                            showSnackbar(getString(R.string.give_permission))
-                        }
-                        updateMap()
-                    }
-                requestLocationPermissionLauncher.launch(android.Manifest.permission.ACCESS_FINE_LOCATION)
-            }
-        }
-//    }
-
-    @SuppressLint("MissingPermission")
-    private fun addBeeAtLocation() {
-        if (map == null) { return }
-        if (fusedLocationProvider == null) { return }
-        if (!locationPermissionGranted) {
-            showSnackbar(getString(R.string.give_permission))
-            return
-        }
-
-        // implement view model and come back to line 124
-
-        fusedLocationProvider?.lastLocation?.addOnCompleteListener(requireActivity()) { locationRequestTask ->
-            val location = locationRequestTask.result
-            if (location != null) {
-                val bee = Bee(
-                    sightingID = UUID.randomUUID(),
-                    numberBees = beeUserSightingNumInput,
-                    location = GeoPoint(location.latitude, location.longitude),
-                    imageRef = 1
-                )
-                beeViewModel.addBee(bee)
-                moveMapToUserLocation()
-                showSnackbar(getString(R.string.bee_sighting_added))
-            } else {
-                showSnackbar(getString(R.string.no_location))
-            }
-        }
-    }
-
-    private fun showSnackbar(message: String) {
-        Snackbar.make(requireView(), message, Snackbar.LENGTH_LONG).show()
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -181,8 +61,14 @@ class BeeSightingUserInput : Fragment() {
 
         val view = inflater.inflate(R.layout.fragment_bee_sighting_user_input, container, false)
 
-        val latitude: Double = requireArguments().getDouble(LATITUDE)
-        val longitude: Double = requireArguments().getDouble(LONGITUDE)
+        val userLocationLat: Double = requireArguments().getDouble(LAT)
+        val userlocationLong: Double = requireArguments().getDouble(LONG)
+        val newUserLocation: GeoPoint = GeoPoint(userLocationLat, userlocationLong)
+
+
+
+
+
 
         increaseButton = view.findViewById(R.id.increaseButton)
         decreaseButton = view.findViewById(R.id.decreaseButton)
@@ -206,34 +92,44 @@ class BeeSightingUserInput : Fragment() {
 
 
         getLocationButton.setOnClickListener {
-            //requestLocationPermission()
-            moveMapToUserLocation()
+            parentFragmentManager.beginTransaction().replace(R.id.bee_fragment_container, GetLocation.newInstance(), "GETLOCATION").commit()
 
         }
 
-        val mapFragment = childFragmentManager.findFragmentById(R.id.map_view) as SupportMapFragment
-        mapFragment?.getMapAsync(mapReadyCallback)
+
 
     // holding code to be replaced with send data to view model - linked to repository - upload to database
         submitButton.setOnClickListener {
-            addBeeAtLocation()
+            val bee = Bee(
+                sightingID = UUID.randomUUID(),
+                numberBees = beeUserSightingNumInput,
+                dateSpotted = Date(),
+                location = newUserLocation,
+                imageRef = 0
 
-            /*parentFragmentManager
-                .beginTransaction()
-                .replace(R.id.bee_fragment_container, BeeBootScreen.newInstance(), "BEEBOOTSCREEN").addToBackStack("Beebootscreen")
-                .commit()*/
+            )
+            beeViewModel.addBee(bee)
 
-        requestLocationPermission()
+            Toast.makeText(context, "Thanks for submitting your data in the bee study!", Toast.LENGTH_LONG).show()
+
+            Timer().schedule(2000) {
+                parentFragmentManager.beginTransaction().replace(
+                    R.id.bee_fragment_container,
+                    BeeBootScreen.newInstance(),
+                    "SubmittedData"
+                ).commit()
+            }
+
         }
         return view
     }
 
     companion object {
         @JvmStatic
-        fun newInstance(latitude:Double?, longitude: Double?) = BeeSightingUserInput().apply {
+        fun newInstance(latitude: Double, longitude: Double) = BeeSightingUserInput().apply {
             arguments = Bundle().apply {
-                putDouble(LATITUDE, latitude!!)
-                putDouble(LONGITUDE, longitude!!)
+                putDouble(LAT, latitude)
+                putDouble(LONG, longitude)
             }
         }
     }
