+++ app/src/main/res/drawable-anydpi/ic_action_location.xml
@@ -1,0 +1,16 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24"
+    android:tint="#333333"
+    android:alpha="0.6">
+  <group android:scaleX="1.2"
+      android:scaleY="1.2"
+      android:translateX="-2.4"
+      android:translateY="-2.4">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M13,6v3h3v2h-3v3h-2v-3H8V9h3V6H13zM18,10.2C18,6.57 15.35,4 12,4s-6,2.57 -6,6.2c0,2.34 1.95,5.44 6,9.14C16.05,15.64 18,12.54 18,10.2zM12,2c4.2,0 8,3.22 8,8.2c0,3.32 -2.67,7.25 -8,11.8c-5.33,-4.55 -8,-8.48 -8,-11.8C4,5.22 7.8,2 12,2z"/>
+  </group>
+</vector>
