+++ app/src/main/res/layout/fragment_bee_sighting_user_input.xml
@@ -54,18 +54,17 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/increaseButton" />
 
-    <ImageView
-        android:id="@+id/imageView3"
-        android:layout_width="205dp"
-        android:layout_height="165dp"
+    <ImageButton
+        android:id="@+id/imageButton"
+        android:layout_width="255dp"
+        android:layout_height="195dp"
         android:layout_marginTop="40dp"
         android:contentDescription="@string/image_view"
-        android:src="@android:drawable/screen_background_light"
+        app:srcCompat="@drawable/ic_action_camera"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/getUserLocationButton" />
-
 
 
     <Button
@@ -74,6 +73,7 @@
         android:layout_height="wrap_content"
         android:layout_margin="40dp"
         android:layout_marginTop="60dp"
+        android:drawableEnd="@drawable/ic_action_location"
         android:text="@string/get_location"
         android:textAlignment="center"
         android:textSize="18sp"
@@ -88,12 +88,13 @@
         android:layout_height="wrap_content"
         android:layout_margin="40dp"
         android:layout_marginTop="40dp"
+        android:drawableEnd="@drawable/ic_action_clouddone"
         android:text="@string/submit"
         android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.495"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageView3" />
+        app:layout_constraintTop_toBottomOf="@+id/imageButton" />
 
     <EditText
         android:id="@+id/beeNumber"
