+++ app/src/main/res/layout/get_location_screen.xml
@@ -2,7 +2,6 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".GetLocation">
@@ -12,6 +11,7 @@
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:backgroundTint="@color/cardview_dark_background"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -20,15 +20,33 @@
         app:layout_constraintVertical_bias="0.0" />
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/floatingActionButton"
+        android:id="@+id/setLocationButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_margin="32dp"
         android:layout_marginBottom="32dp"
         android:clickable="true"
+        android:paddingTop="2dp"
+        android:src="@drawable/ic_action_name"
         app:layout_constraintBottom_toBottomOf="@+id/map_view"
-        app:layout_constraintStart_toStartOf="@+id/map_view"
-        app:srcCompat="@android:drawable/checkbox_on_background" />
+        app:layout_constraintEnd_toEndOf="@+id/clickInstructions"
+        app:layout_constraintHorizontal_bias="0.513"
+        app:layout_constraintStart_toStartOf="@+id/clickInstructions" />
+
+    <TextView
+        android:id="@+id/clickInstructions"
+        android:layout_width="132dp"
+        android:layout_height="76dp"
+        android:layout_marginStart="10dp"
+        android:layout_marginBottom="5dp"
+        android:background="@android:color/system_outline_variant_light"
+        android:paddingTop="2dp"
+        android:text="@string/click_the_box_after_setting_location"
+        android:textAlignment="center"
+        android:textSize="19sp"
+        android:textStyle="bold"
+        app:layout_constraintBottom_toTopOf="@+id/setLocationButton"
+        app:layout_constraintStart_toStartOf="@+id/map_view" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
